:truck: rename dcc_page
